b.root pdns_recursor speedups mysql support for ports again valgrind fixes
git-svn-id: svn://svn.powerdns.com/pdns/trunk/pdns/pdns@228 d19b8d6e-7fed-0310-83ef-9ca221ded41b

+++ syncres.cc
@@ -332,7 +332,7 @@ inline vector<string> SyncRes::shuffle(set<string> &nameservers, const string &p
 	if(!((i-rnameservers.begin())%4))
 	  L<<endl<<Logger::Warning<<prefix<<"             ";
       }
-      L<<*i<<"(" << speeds[*i]/1000.0 <<"ms)";
+      L<<*i<<"(" << (int)(speeds[*i]/1000.0) <<"ms)";
     }
     L<<endl;
   }
